Changes on way to JWT authorization

+++ src/main/java/com/pozdeev/HelloWorld/repositories/CommentRepo.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CommentRepo extends CrudRepository<Comment, Integer> {
+public interface CommentRepo extends CrudRepository<Comment, Long> {
 
     List<Comment> findAllByOrderByCommentIdDesc();
 }
