+++ src/main/java/com/pozdeev/HelloWorld/repositories/UserRepo.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepo extends CrudRepository<User, Integer> {
+public interface UserRepo extends CrudRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
